encoder: fix duration handling in VBV lookahead.

+++ source/common/frame.h
@@ -183,13 +183,13 @@ public:
     int                    m_targetQp;
 
     /*HRD delays for this frame*/
-    int64_t                m_duration;           // display duration of picture in ticks (VUI timebase)
-    float                  m_timebase;           // the timebase for this picture [seconds]
-    int64_t                m_plannedCpbDuration; // lifetime in the CPB (VUI timebase)
-    int64_t                m_cpbDelay;           // removal delay of the picture (VUI timebase)
-    uint64_t               m_displayPicCount;    // display picture count
-    uint64_t               m_codedPicCount;      // remove me, for debugging
-    int64_t                m_dpbOutputDelay;     // output delay of the picture (VUI timebase)
+    unsigned int           m_duration;           // display duration of picture in ticks (VUI timebase)
+    double                 m_timebase;           // the timebase for this picture [seconds]
+    unsigned int           m_plannedCpbDuration; // lifetime in the CPB (VUI timebase)
+    unsigned int           m_cpbDelay;           // removal delay of the picture (VUI timebase)
+    uint64_t               m_displayPicCount;    // display picture count (counted in decoder ticks)
+    uint64_t               m_codedPicCount;      // coded picture count (counted in decoder ticks)
+    unsigned int           m_dpbOutputDelay;     // output delay of the picture (VUI timebase)
 
     Frame();
